+++ DalaMock.PluginTemplate/template/SamplePlugin/Windows/ConfigWindow.cs
@@ -1,7 +1,7 @@
 namespace SamplePlugin.Windows;
 
 using Dalamud.Interface.Windowing;
-using ImGuiNET;
+using Dalamud.Bindings.ImGui;
 
 public class ConfigWindow(Configuration config) : Window("SamplePlugin Config")
 {