+++ DalaMock.PluginTemplate/template/SamplePlugin/Windows/MainWindow.cs
@@ -4,7 +4,7 @@ using DalaMock.Shared.Interfaces;
 using Dalamud.Interface;
 using Dalamud.Interface.Utility.Raii;
 using Dalamud.Interface.Windowing;
-using ImGuiNET;
+using Dalamud.Bindings.ImGui;
 
 public class MainWindow : Window
 {